requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
pyairtable==2.1.0
requests==2.31.0
python-jwt==4.0.0
pyjwt==2.8.0
bcrypt==4.0.1
